Add files via upload
The insurance case study was a data analysis using Python in Jupyter Lab to analyze the overall behavior of the customers while claiming their claims, post the lost of their properties. This includes cleaning the data, the amount claimed for each type of accidents and utilizing the statistical tools of python libraries to analyze the relationship between the customers and the insurance claims, their age and the segment and the relationship with the claim amounts and the gender.